newtekreviews/requirements.txt
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
comm==0.2.2
cryptography==43.0.3
debugpy==1.8.7
decorator==5.1.1
defusedxml==0.8.0rc2
Django==5.0.6
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-spectacular==0.27.2
executing==2.0.1
hiredis==3.0.0
idna==3.10
inflection==0.5.1
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
oauthlib==3.2.2
packaging==24.1
parso==0.8.4
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.47
psutil==6.0.0
psycopg2==2.9.10
pure-eval==0.2.2
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pywin32==307
PyYAML==6.0.2
pyzmq==26.2.0
redis==5.1.1
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.20.0
setuptools==71.0.3
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
stack-data==0.6.3
tornado==6.4.1
traitlets==5.14.3
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
wcwidth==0.2.13
